+++ CHANGELOG.md
@@ -2,9 +2,11 @@ Important changes to data-models, configuration and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bumped runtimeVersion to `2021.08.16`.
+ * NOTE: started to populate and use `PackageDocument.uploaderUserIds`
+         TODO(deferred): remove `PackageDocument.uploaderEmails` in a future release
 
 ## `20210812t143400-all`
-
  * Bumped runtimeVersion to `2021.08.12`.
  * Upgraded preview Dart analysis SDK to `2.14.0-377.4.beta`
  * Upgraded preview Flutter analysis SDK to `2.5.0-5.1.pre`.